requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
click==8.3.1
colorama==0.4.6
fastapi==0.128.0
greenlet==3.3.0
h11==0.16.0
idna==3.11
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pytesseract==0.3.13
python-multipart==0.0.21
SQLAlchemy==2.0.45
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
PyMuPDF==1.24.14
